Fix standard regular expressions to skip headers
So we had an issue with the below output as it would repeat the headers as part of the returned values.

The solution is to fix the regular expressions, as they would match too broad. For instance the expression `(\w+.\w+.\w+)` would match things like `A B C` or `Nightmare|On|ElmStreet`. Dots should be matched using `\.` and for a mac-address the regexp could be much more strict.

Same for vlan numbers we know it's between 1 and 4 digits.