Made tox.ini work for ntc-templates
The previous version didn't work, and tried to bring in
ansible and a variety of other things that looks like it was a mishmash
of travis-ci and ansible testing. These templates do not require or
use ansible during testing, and they should test the version of
ntc-templates in your workspace, not pull some arbitrary version from
some arbitrary repo and test that version.

Ansible testing is the domain of ansible-related repos, not this one.
git commands are the domain of travis-ci, not tox.

This version of tox passes tests, and tests against Python 3.7 as well
as 2.7.