Bug Fix: Cisco IOS show mac address-table
Fixed a bug with the cisco_ios_show_mac-address-table.textfsm template
where it error on lines that had a '*' and and 'R' on the same line at
the beginning of the line. See the new test files for an example.

I encountered this in the wild on some Cisco IOS devices.