Updated huawei_smartax display ont info \S+ template (#1704)
so it matches other versions of the template.

In certain edge cases was not working. For example
with the model EA5608 was not matching correctly,
because the output is slightly different. Now it
works!

Also added the noqa E231 exception in tasks.py as
it is not passing the flake8 tests. Now they pass.
The error is due to a detection of the : in side
a string. It shouldn't be detected in that case
but flake8 is detecting it.