Add Linux iwconfig (#2122)
* Adding new template for Linux: `iwconfig`

* Change command regex in index

* Define recording regex beginning with "wl"

Further define the recording regex beginning with "wl" so it is
less likely to inadvertently match other text.

Examples it will match:

- wlan0
- wlp3s0

* Swap whitespace regex for no wireless extensions

Swap \s+ for \s* where there is at least one space.
Doing so for flexibility reasons.

* Remove whitespace regex after Enc key:

Since none of the labels have a space before the colon it doesn't
seem necessary to have the extra characters.

* removing the 802.11 Mode

---------

Co-authored-by: Michael Bear <38406045+mjbear@users.noreply.github.com>